Change: マイページのルーティングを単一リソースに切り出し
