+++ app/controllers/my_pages_controller.rb
@@ -1,0 +1,26 @@
+class MyPagesController < ApplicationController
+  before_action :set_current_user, only: %i[edit update]
+
+  def show; end
+
+  def edit; end
+
+  def update
+    if @user.update(user_params)
+      redirect_to my_page_path, success: t('.success')
+    else
+      flash.now[:danger] = t('.failed')
+      render :edit
+    end
+  end
+
+  private
+
+  def set_current_user
+    @user = User.find(current_user.id)
+  end
+
+  def user_params
+    params.require(:user).permit(:name, :email, :password, :password_confirmation)
+  end
+end
