+++ app/controllers/users_controller.rb
@@ -1,5 +1,4 @@
 class UsersController < ApplicationController
-  before_action :set_current_user, only: %i[edit update]
   skip_before_action :require_login, only: %i[new create], raise: false
 
   # 遠足のおやつは300円まで
@@ -8,8 +7,6 @@ class UsersController < ApplicationController
   def index
     @users = User.all
   end
-
-  def show; end
 
   def new
     @user = User.new
@@ -27,23 +24,7 @@ class UsersController < ApplicationController
     end
   end
 
-  def edit; end
-
-  def update
-    if @user.update(user_params)
-      redirect_to user_path(current_user.id), success: t('.success')
-    else
-      binding.pry
-      flash.now[:danger] = 'しっぱい'
-      render :edit
-    end
-  end
-
   private
-
-  def set_current_user
-    @user = User.find(current_user.id)
-  end
 
   def user_params
     params.require(:user).permit(:name, :email, :password, :password_confirmation)
