+++ app/helpers/my_pages_helper.rb
@@ -1,0 +1,2 @@
+module MyPagesHelper
+end
