+++ app/views/my_pages/show.html.slim
@@ -11,4 +11,4 @@ div.main-wrapper
       div.column.mb-3
         h2 めーるあどれす:
         h3= "#{current_user.email}"
-      = link_to 'こうしんする', edit_user_path(current_user.id), class: 'btn btn-outline-light'
+      = link_to 'こうしんする', edit_my_page_path, class: 'btn btn-outline-light'
