+++ config/routes.rb
@@ -8,7 +8,8 @@ Rails.application.routes.draw do
   get '/choose_oyatsu', to: 'oyatsus#index'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 
-  resources :users do
+  resources :users, only: %i[index new create] do
     resources :baskets, only: %i[index show new create destroy], shallow: true
   end
+  resource :my_page, only: %i[show edit update]
 end
